+++ src/system/ast/Send.c
@@ -23,7 +23,7 @@ Send new_Send(Optr receiver, Optr msg, uns_int argc, ...)
     Send result = new_Send_raw(receiver, msg, argc);
     va_list args;
     va_start(args, argc);
-    int idx;
+    long idx;
     for (idx = 0; idx < argc; idx++) {
         Send_args(result)[idx] = va_arg(args, Optr);
     }
