+++ src/system/collection/Array.h
@@ -18,12 +18,12 @@ extern Array new_Array_raw(uns_int c);
 extern Array new_Array(uns_int c, Optr v[]);
 extern Array new_Array_with(uns_int c, ...);
 extern Array new_Array_withAll(uns_int c, Optr element);
-extern Optr raw_Array_at(Array array, Optr tag, int index);
+extern Optr raw_Array_at(Array array, Optr tag, long index);
 extern void raw_Array_at_put(Array array, Optr tag,
-                                  int index, Optr value);
-extern Optr raw_Array_instAt(Array array, Optr tag, int index);
+                                  long index, Optr value);
+extern Optr raw_Array_instAt(Array array, Optr tag, long index);
 extern void raw_Array_instAt_put(Array array, Optr tag,
-                                      int index, Optr value);
+                                      long index, Optr value);
 /* ========================================================================= */
 
 #endif // ARRAY_H
